requirements.txt
streamlit>=1.40.0
anthropic>=0.40.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
python-docx>=1.1.0
